+++ examples/custom_routes/lib/pwa/worker.dart
@@ -6,7 +6,7 @@ Worker createWorker() {
   Worker worker = new Worker()..offlineUrls = offlineUrls;
 
   DynamicCache youtubeThumbnails =
-      new DynamicCache('youtube', maxEntries: 10, noNetworkCaching: true);
+      new DynamicCache('youtube', maxEntries: 10, skipDiskCache: true);
 
   worker.router.get('https://i.ytimg.com/vi/', youtubeThumbnails.cacheFirst);
   return worker;